data_visualization_learning
Data visualization is the process of converting raw data into meaningful visual representations. It is an important aspect of data analysis because it helps to understand the data better, spot patterns and hidden insights, and present findings to others in a way that is easy to understand. By visualizing the data, we can process information much more efficiently than by simply looking at raw numbers. Furthermore, with well-crafted charts and dashboards, data analysts can communicate their insights to people with varying levels of domain expertise. Data visualization has the power to bridge the gap between data analysts and others, such as business executives, supervisors, project managers, and clients, by presenting information in a format that is easily understood by all